+++ startup/gui/ocio.py
@@ -35,6 +35,7 @@
 #
 ##########################################################################
 
+import copy
 import functools
 import imath
 
@@ -76,16 +77,14 @@ Gaffer.Metadata.registerValue( preferences["displayColorSpace"]["context"], "lay
 
 # Register with `GafferUI.DisplayTransform` for use by Widgets.
 
-def __processor( view ) :
+def __processor( display, view ) :
 
 	d = OCIO.DisplayViewTransform()
 	d.setSrc( OCIO.ROLE_SCENE_LINEAR )
-	d.setDisplay( defaultDisplay )
+	d.setDisplay( display )
 	d.setView( view )
 
-	# \todo : Should be `context = copy.deepcopy( config.getCurrentContext() )`
-	# once https://github.com/AcademySoftwareFoundation/OpenColorIO/pull/1575 gets merged into OCIO2.1.2
-	context = OCIO.Context( searchPaths = list( config.getCurrentContext().getSearchPaths() ), workingDir = config.getCurrentContext().getWorkingDir(), environmentMode = config.getCurrentContext().getEnvironmentMode(), stringVars = dict( config.getCurrentContext().getStringVars() ) )
+	context = copy.deepcopy( config.getCurrentContext() )
 	gafferContext = Gaffer.Context.current()
 	for variable in preferences["displayColorSpace"]["context"] :
 		if variable["enabled"].getValue() :
@@ -95,7 +94,7 @@ def __processor( view ) :
 
 def __setDisplayTransform() :
 
-	cpuProcessor = __processor( preferences["displayColorSpace"]["view"].getValue() ).getDefaultCPUProcessor()
+	cpuProcessor = __processor( defaultDisplay, preferences["displayColorSpace"]["view"].getValue() ).getDefaultCPUProcessor()
 
 	def f( c ) :
 
@@ -106,41 +105,113 @@ def __setDisplayTransform() :
 
 __setDisplayTransform()
 
-# Register with `GafferUI.View.DisplayTransform` for use in the Viewer.
-
-def __displayTransformCreator( view ) :
-
-	processor = __processor( view )
-	return GafferImageUI.OpenColorIOAlgo.displayTransformToFramebufferShader( processor )
-
-def __registerViewerDisplayTransforms() :
-
-	for name in config.getViews( defaultDisplay ) :
-		GafferUI.View.DisplayTransform.registerDisplayTransform(
-			name,
-			functools.partial( __displayTransformCreator, name )
-		)
-
-	GafferUI.View.DisplayTransform.registerDisplayTransform(
-		"Default",
-		functools.partial( __displayTransformCreator, preferences["displayColorSpace"]["view"].getValue() )
-	)
-
-__registerViewerDisplayTransforms()
-
-# And connect to `plugSet()` to update everything again when the user modifies something.
+# And connect to `plugSet()` to update `GafferUI.DisplayTransform` again when the user modifies something.
 
 def __plugSet( plug ) :
 
 	if plug.relativeName( plug.node() ) != "displayColorSpace" :
 		return
 
-	__registerViewerDisplayTransforms()
 	__setDisplayTransform()
 
 preferences.plugSetSignal().connect( __plugSet, scoped = False )
 
-Gaffer.Metadata.registerValue( GafferUI.View, "displayTransform.name", "userDefault", "Default" )
+# Register with `GafferUI.View.DisplayTransform` for use in the Viewer.
+
+def __displayTransformCreator( display, view ) :
+
+	processor = __processor( display, view )
+	return GafferImageUI.OpenColorIOAlgo.displayTransformToFramebufferShader( processor )
+
+def __registerViewerDisplayTransforms() :
+
+	for display in config.getDisplays() :
+		for view in config.getViews( display ) :
+			GafferUI.View.DisplayTransform.registerDisplayTransform(
+				f"{display}/{view}",
+				functools.partial( __displayTransformCreator, display, view )
+			)
+
+__registerViewerDisplayTransforms()
+
+class DisplayTransformPlugValueWidget( GafferUI.PlugValueWidget ) :
+
+	def __init__( self, plugs, **kw ) :
+
+		self.__menuButton = GafferUI.MenuButton( "", menu = GafferUI.Menu( Gaffer.WeakMethod( self.__menuDefinition ) ) )
+		GafferUI.PlugValueWidget.__init__( self, self.__menuButton, plugs, **kw )
+
+		self.__currentValue = ""
+
+	def _updateFromValues( self, values, exception ) :
+
+		if exception is not None :
+			self.__menuButton.setText( "" )
+			self.__currentValue = ""
+		else :
+			assert( len( values ) == 1 )
+			self.__currentValue = values[0]
+			# Only show the View name, because the Display name is more of
+			# a "set once and forget" affair. The menu shows both for when
+			# you need to check.
+			self.__menuButton.setText( self.__currentValue.partition( "/" )[-1] )
+
+		self.__menuButton.setErrored( exception is not None )
+
+	def _updateFromEditable( self ) :
+
+		self.__menuButton.setEnabled( self._editable() )
+
+	def __menuDefinition( self ) :
+
+		result = IECore.MenuDefinition()
+
+		activeViews = Gaffer.Metadata.value( self.getPlug(), "openColorIO:activeViews" ) or "*"
+
+		# View section
+
+		result.append( "/__ViewDivider__", { "divider" : True, "label" : "View" } )
+
+		displayToViews = {}
+
+		currentDisplay, currentView = self.__currentValue.split( "/" )
+		for displayTransform in GafferUI.View.DisplayTransform.registeredDisplayTransforms() :
+			display, view = displayTransform.split( "/" )
+			if not IECore.StringAlgo.matchMultiple( view, activeViews ) :
+				continue
+			displayToViews.setdefault( display, [] ).append( view )
+			if display != currentDisplay :
+				continue
+			result.append(
+				f"/{view}", {
+					"command" : functools.partial( Gaffer.WeakMethod( self.__setValue ), f"{currentDisplay}/{view}" ),
+					"checkBox" : view == currentView
+				}
+			)
+
+		# Display section
+
+		result.append( "/__DisplayDivider__", { "divider" : True, "label" : "Display" } )
+
+		for display, views in displayToViews.items() :
+			newValue = "{}/{}".format( display, currentView if currentView in views else views[0] )
+			result.append(
+				f"/{display}", {
+					"command" : functools.partial( Gaffer.WeakMethod( self.__setValue ), newValue ),
+					"checkBox" : display == currentDisplay
+				}
+			)
+
+		return result
+
+	def __setValue( self, value, unused ) :
+
+		self.getPlug().setValue( value )
+
+GafferImageUI._OpenColorIODisplayTransformPlugValueWidget = DisplayTransformPlugValueWidget
+Gaffer.Metadata.registerValue( GafferUI.View, "displayTransform.name", "plugValueWidget:type", "GafferImageUI._OpenColorIODisplayTransformPlugValueWidget" )
+Gaffer.Metadata.registerValue( GafferUI.View, "displayTransform.name", "userDefault", "{}/{}".format( defaultDisplay, config.getDefaultView( defaultDisplay ) ) )
+Gaffer.Metadata.registerValue( GafferUI.View, "displayTransform.name", "layout:minimumWidth", 150 )
 
 # Add "Roles" submenus to various colorspace plugs. The OCIO UX guidelines suggest we
 # shouldn't do this, but they do seem like they might be useful, and historically they